pyproject.toml
[tool.poetry]
name = "ood-testing"
version = "0.1.0"
description = ""
authors = ["martilut <jeykolton@gmail.com>"]
readme = "README.md"
packages = [{ include = "oodt" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"

# Core numerical stack
numpy = "^1.26"
scipy = "^1.11"
pandas = "^2.1"

# ML core
scikit-learn = "^1.4"

# Tree-based tabular models (common baselines)
xgboost = { version = "^2.0", optional = true }
catboost = { version = "^1.2", optional = true }

# Metrics and statistics
statsmodels = "^0.14"

# Visualization (for analysis, not core runtime)
matplotlib = "^3.8"
seaborn = { version = "^0.13", optional = true }

# Progress & logging
tqdm = "^4.66"
hydra-core = "^1.3.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
black = "^24.1"
mypy = "^1.8"

[tool.poetry.extras]
trees = ["xgboost", "catboost"]
viz = ["seaborn"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
